move predicate package and merge with predicate utils
