+++ core/evm.go
@@ -32,8 +32,7 @@ import (
 	"github.com/ava-labs/subnet-evm/consensus"
 	"github.com/ava-labs/subnet-evm/core/types"
 	"github.com/ava-labs/subnet-evm/core/vm"
-	"github.com/ava-labs/subnet-evm/precompile/results"
-	"github.com/ava-labs/subnet-evm/utils/predicate"
+	"github.com/ava-labs/subnet-evm/predicate"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/log"
 	//"github.com/ethereum/go-ethereum/log"
@@ -51,9 +50,9 @@ type ChainContext interface {
 
 // NewEVMBlockContext creates a new context for use in the EVM.
 func NewEVMBlockContext(header *types.Header, chain ChainContext, author *common.Address) vm.BlockContext {
-	var predicateResults *results.PredicateResults
+	var predicateResults *predicate.PredicateResults
 	if predicateBytes, err := predicate.GetPredicateResultBytes(header.Extra); err == nil {
-		predicateResults, err = results.ParsePredicateResults(predicateBytes)
+		predicateResults, err = predicate.ParsePredicateResults(predicateBytes)
 		if err != nil {
 			log.Error("failed to parse predicate results creating new block context", "err", err, "extra", header.Extra)
 		}
@@ -66,11 +65,11 @@ func NewEVMBlockContext(header *types.Header, chain ChainContext, author *common
 // in header.Extra.
 // This function is used to create a BlockContext when the header Extra data is not fully formed yet and it's more efficient to pass in predicateResults
 // directly rather than re-encode the latest results when executing each individaul transaction.
-func NewEVMBlockContextWithPredicateResults(header *types.Header, chain ChainContext, author *common.Address, predicateResults *results.PredicateResults) vm.BlockContext {
+func NewEVMBlockContextWithPredicateResults(header *types.Header, chain ChainContext, author *common.Address, predicateResults *predicate.PredicateResults) vm.BlockContext {
 	return newEVMBlockContext(header, chain, author, predicateResults)
 }
 
-func newEVMBlockContext(header *types.Header, chain ChainContext, author *common.Address, predicateResults *results.PredicateResults) vm.BlockContext {
+func newEVMBlockContext(header *types.Header, chain ChainContext, author *common.Address, predicateResults *predicate.PredicateResults) vm.BlockContext {
 	var (
 		beneficiary common.Address
 		baseFee     *big.Int
