+++ core/vm/evm.go
@@ -38,7 +38,7 @@ import (
 	"github.com/ava-labs/subnet-evm/precompile/contracts/deployerallowlist"
 	"github.com/ava-labs/subnet-evm/precompile/modules"
 	"github.com/ava-labs/subnet-evm/precompile/precompileconfig"
-	"github.com/ava-labs/subnet-evm/precompile/results"
+	"github.com/ava-labs/subnet-evm/predicate"
 	"github.com/ava-labs/subnet-evm/vmerrs"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/crypto"
@@ -114,7 +114,7 @@ type BlockContext struct {
 	GetHash GetHashFunc
 	// PredicateResults are the results of predicate verification available throughout the EVM's execution.
 	// PredicateResults may be nil if it is not encoded in the block's header.
-	PredicateResults *results.PredicateResults
+	PredicateResults *predicate.PredicateResults
 
 	// Block information
 	Coinbase    common.Address // Provides information for COINBASE
