+++ plugin/evm/block.go
@@ -18,8 +18,8 @@ import (
 	"github.com/ava-labs/subnet-evm/core/types"
 	"github.com/ava-labs/subnet-evm/params"
 	"github.com/ava-labs/subnet-evm/precompile/precompileconfig"
-	"github.com/ava-labs/subnet-evm/precompile/results"
-	"github.com/ava-labs/subnet-evm/utils/predicate"
+	"github.com/ava-labs/subnet-evm/predicate"
+	predicate1 "github.com/ava-labs/subnet-evm/predicate"
 	"github.com/ava-labs/subnet-evm/warp/payload"
 	"github.com/ava-labs/subnet-evm/x/warp"
 
@@ -276,7 +276,7 @@ func (b *Block) verifyPredicates(predicateContext *precompileconfig.PredicateCon
 		return nil
 	}
 
-	predicateResults := results.NewPredicateResults()
+	predicateResults := predicate1.NewPredicateResults()
 	for _, tx := range b.ethBlock.Transactions() {
 		results, err := core.CheckPredicates(rules, predicateContext, tx)
 		if err != nil {
@@ -295,7 +295,7 @@ func (b *Block) verifyPredicates(predicateContext *precompileconfig.PredicateCon
 		return err
 	}
 	if !bytes.Equal(headerPredicateResultsBytes, predicateResultsBytes) {
-		parsedResults, err := results.ParsePredicateResults(headerPredicateResultsBytes)
+		parsedResults, err := predicate1.ParsePredicateResults(headerPredicateResultsBytes)
 		if err != nil {
 			return err
 		}
