+++ plugin/evm/syncervm_test.go
@@ -32,7 +32,7 @@ import (
 	"github.com/ava-labs/subnet-evm/ethdb"
 	"github.com/ava-labs/subnet-evm/metrics"
 	"github.com/ava-labs/subnet-evm/params"
-	"github.com/ava-labs/subnet-evm/precompile/results"
+	"github.com/ava-labs/subnet-evm/predicate"
 	statesyncclient "github.com/ava-labs/subnet-evm/sync/client"
 	"github.com/ava-labs/subnet-evm/sync/statesync"
 	"github.com/ava-labs/subnet-evm/trie"
@@ -264,9 +264,7 @@ func TestVMShutdownWhileSyncing(t *testing.T) {
 }
 
 func createSyncServerAndClientVMs(t *testing.T, test syncTest) *syncVMSetup {
-	var (
-		serverVM, syncerVM *VM
-	)
+	var serverVM, syncerVM *VM
 	// If there is an error shutdown the VMs if they have been instantiated
 	defer func() {
 		// If the test has not already failed, shut down the VMs since the caller
@@ -293,7 +291,7 @@ func createSyncServerAndClientVMs(t *testing.T, test syncTest) *syncVMSetup {
 	// configure [serverVM]
 	_, serverVM, _, serverAppSender := GenesisVM(t, true, genesisJSONLatest, "", "")
 	generateAndAcceptBlocks(t, serverVM, parentsToGet, func(i int, gen *core.BlockGen) {
-		b, err := results.NewPredicateResults().Bytes()
+		b, err := predicate.NewPredicateResults().Bytes()
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -469,7 +467,7 @@ func testSyncerVM(t *testing.T, vmSetup *syncVMSetup, test syncTest) {
 	txsPerBlock := 10
 	toAddress := testEthAddrs[1] // arbitrary choice
 	generateAndAcceptBlocks(t, syncerVM, blocksToBuild, func(_ int, gen *core.BlockGen) {
-		b, err := results.NewPredicateResults().Bytes()
+		b, err := predicate.NewPredicateResults().Bytes()
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -495,7 +493,7 @@ func testSyncerVM(t *testing.T, vmSetup *syncVMSetup, test syncTest) {
 
 	// Generate blocks after we have entered normal consensus as well
 	generateAndAcceptBlocks(t, syncerVM, blocksToBuild, func(_ int, gen *core.BlockGen) {
-		b, err := results.NewPredicateResults().Bytes()
+		b, err := predicate.NewPredicateResults().Bytes()
 		if err != nil {
 			t.Fatal(err)
 		}
