+++ predicate/predicate_bytes.go
