+++ predicate/predicate_bytes_test.go
